UI 3

+++ propertyManagement/src/model/relation/PropertyRecord.java
@@ -32,7 +32,7 @@ public class PropertyRecord {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
